moved components to wrapper and addressed other
